@@ -1,20 +1,20 @@
 source 'https://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.0.2'
+gem 'rails', '4.2.0'
 
 # Use sqlite3 as the database for Active Record
 #gem 'sqlite3'
 gem 'pg'
 
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 4.0.0'
+gem 'sass-rails'
 
 # Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '>= 1.3.0'
+gem 'uglifier'
 
 # Use CoffeeScript for .js.coffee assets and views
-gem 'coffee-rails', '~> 4.0.0'
+gem 'coffee-rails'
 
 # See https://github.com/sstephenson/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
@@ -26,7 +26,7 @@ gem 'jquery-rails'
 gem 'turbolinks'
 
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 1.2'
+gem 'jbuilder'
 
 group :doc do
   # bundle exec rake doc:rails generates the API under doc/api.
@@ -34,7 +34,7 @@ group :doc do
 end
 
 # Use ActiveModel has_secure_password
-# gem 'bcrypt-ruby', '~> 3.1.2'
+gem 'bcrypt', require: "bcrypt"
 
 # Use unicorn as the app server
 # gem 'unicorn'
@@ -45,16 +45,17 @@ end
 # Use debugger
 # gem 'debugger', group: [:development, :test]
 
-gem 'devise'
+gem 'foundation-rails'
+gem 'foundation_rails_helper'
 
 # Used to make printable pdf transaction records
 # This gem was highly recommended by railscast
 # prawnpdf.org/manual.pdf‎
 # set to stable release, for no future problems.
-gem 'prawn',   '= 1.0.0'
+gem 'prawn'
 # makes using prawn easier by using the controllers show
 # no added routes
-gem 'prawnto', '= 0.1.1'
+gem 'prawnto'
 
 # Automatically run scripts
 gem 'whenever', :require => false
