@@ -1,60 +1,21 @@
-class SessionsController  < Devise::SessionsController
-  prepend_before_filter :require_no_authentication, only: [ :new, :create ]
-  prepend_before_filter :allow_params_authentication!, only: :create
-  prepend_before_filter only: [ :create, :destroy ]# { request.env["devise.skip_timeout"] = true }
-  # GET /resource/sign_in
+class SessionsController < ApplicationController
+  skip_before_action :require_login, only: [:new, :create]
   def new
-    self.resource = resource_class.new(sign_in_params)
-    clean_up_passwords(resource)
-    respond_with(resource, serialize_options(resource))
   end
 
-  # POST /resource/sign_in
   def create
-    self.resource = warden.authenticate!(auth_options)
-    set_flash_message(:notice, :signed_in) if is_flashing_format?
-    sign_in(resource_name, resource)
-    yield resource if block_given?
-    if admin?
-      respond_with resource, location: admin_panel_path
-    elsif teller?
-      respond_with resource, location: teller_panel_path
-    elsif customer?
-      respond_with resource, location: customer_path
+    user = User.find_by_email(params[:email])
+    if user && user.authenticate(params[:password])
+      session[:user_id] = user.id
+      redirect_to root_url, :notice => "Logged in!"
     else
-      respond_with resource, location: after_sign_in_path_for(resource)
+      flash.now.alert = "Invalid email or password"
+      render "new"
     end
   end
 
-  # DELETE /resource/sign_out
   def destroy
-    redirect_path = after_sign_out_path_for(resource_name)
-    signed_out = (Devise.sign_out_all_scopes ? sign_out : sign_out(resource_name))
-    set_flash_message :notice, :signed_out if signed_out && is_flashing_format?
-    yield if block_given?
-
-    # We actually need to hardcode this as Rails default responder doesn't
-    # support returning empty response on GET request
-    respond_to do |format|
-      format.all { head :no_content }
-      format.any(*navigational_formats) { redirect_to redirect_path }
-    end
-  end
-
-  protected
-
-  def sign_in_params
-    devise_parameter_sanitizer.sanitize(:sign_in)
-  end
-
-  def serialize_options(resource)
-    methods = resource_class.authentication_keys.dup
-    methods = methods.keys if methods.is_a?(Hash)
-    methods << :password if resource.respond_to?(:password)
-    { methods: methods, only: [:password] }
-  end
-
-  def auth_options
-    { scope: resource_name, recall: "#{controller_path}#new" }
+    session[:user_id] = nil
+    redirect_to root_url, :notice => "Logged out!"
   end
 end
