@@ -11,12 +11,12 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20140411034730) do
+ActiveRecord::Schema.define(version: 20140329080945) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "accounts", force: true do |t|
+  create_table "accounts", force: :cascade do |t|
     t.integer  "user_id"
     t.float    "current_balance"
     t.string   "account_type"
@@ -26,7 +26,7 @@ ActiveRecord::Schema.define(version: 20140411034730) do
     t.datetime "updated_at"
   end
 
-  create_table "bills", force: true do |t|
+  create_table "bills", force: :cascade do |t|
     t.integer  "user_id"
     t.integer  "account_id"
     t.boolean  "is_recurring"
@@ -42,7 +42,7 @@ ActiveRecord::Schema.define(version: 20140411034730) do
     t.datetime "updated_at"
   end
 
-  create_table "transactions", force: true do |t|
+  create_table "transactions", force: :cascade do |t|
     t.integer  "user_id"
     t.integer  "account_id"
     t.float    "amount"
@@ -52,24 +52,23 @@ ActiveRecord::Schema.define(version: 20140411034730) do
     t.datetime "updated_at"
   end
 
-  create_table "users", force: true do |t|
-    t.string   "email",                  default: "", null: false
-    t.string   "encrypted_password",     default: "", null: false
+  create_table "users", force: :cascade do |t|
+    t.string   "email"
+    t.string   "username"
+    t.string   "first_name"
+    t.string   "last_name"
+    t.string   "password_digest"
     t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          default: 0,  null: false
+    t.integer  "sign_in_count",          default: 0, null: false
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
     t.string   "last_sign_in_ip"
+    t.integer  "user_level"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "username"
-    t.integer  "user_level"
   end
-
-  add_index "users", ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true, using: :btree
-  add_index "users", ["username"], name: "index_users_on_username", unique: true, using: :btree
 
 end
